자바 코드 받아와서 cmd 이용해서 실행후 결과값 api로 반환

+++ src/main/java/com/example/jwt/refresh/study/jwt/auth/web/AuthController.java
@@ -8,6 +8,7 @@ import com.example.jwt.refresh.study.jwt.boot.util.JwtUtils;
 import com.example.jwt.refresh.study.jwt.example.service.WasmTestService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.python.antlr.ast.While;
 import org.python.util.PythonInterpreter;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -17,6 +18,7 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.*;
+import java.util.Scanner;
 import java.util.UUID;
 
 @Slf4j
@@ -94,12 +96,12 @@ public class AuthController {
 
     @PostMapping("/process")
     public ResponseEntity<String> processss(@RequestBody(required = false) String pyScript) throws IOException, InterruptedException {
-        String randoom = UUID.randomUUID() + ".py";
+        String randoom =  "Test3.java";
+        String randoom2 = randoom.replace(".java", "");
 
         String path2 = "C:\\Users\\wpghk\\Downloads\\jwtrefresh-91d0bc1cde9b278d3d7f7e4c640c9fa57893df1f\\study.jwt\\src\\main\\resources";
-        String filePath = "C:\\Users\\wpghk\\Downloads\\jwtrefresh-91d0bc1cde9b278d3d7f7e4c640c9fa57893df1f\\study.jwt\\src\\main\\resources\\"+randoom;
 
-        File file = new File(filePath);
+        File file = new File(path2 +"\\" + randoom);
         if(!file.exists()) {
             file.createNewFile();
         }
@@ -114,24 +116,50 @@ public class AuthController {
         if(file.exists()) {
             ProcessBuilder builder = new ProcessBuilder();
             builder.directory(new File(path2));
-            builder.command("python", randoom);
-
+            builder.command("javac", randoom);
+            builder.start();
+            builder.command("java", "Test");
             Process process = builder.start();
 
             BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
 
-            StringBuffer line = new StringBuffer();
+            StringBuffer ttt = new StringBuffer();
             String mola = null;
-            while((mola = reader.readLine()) != null) {
-                line.append(mola+"\n");
-                log.info(line.toString());
+            while ((mola = reader.readLine()) != null) {
+                ttt.append(mola);
+                log.info(ttt.toString());
             }
 
-            if(process != null) {
+            if (process != null) {
                 process.destroy();
             }
-            file.delete();
-            return new ResponseEntity<>(line.toString(), HttpStatus.OK);
+
+//            Thread.sleep(3000);
+            while(!(ttt.equals("done"))) {
+
+                if(new File(path2 +"\\" + "Test3.class").exists()){
+                    builder.command("java", "Test3");
+                    Process process2 = builder.start();
+
+                    BufferedReader reader2 = new BufferedReader(new InputStreamReader(process2.getInputStream()));
+
+                    StringBuffer line2 = new StringBuffer();
+                    String mola2 = null;
+                    while ((mola2 = reader2.readLine()) != null) {
+                        line2.append(mola2).append("\n");
+                    }
+
+                    if (process2 != null) {
+                        process2.destroy();
+                    }
+
+                    file.delete();
+                    new File(path2 +"\\" + "Test3.class").delete();
+
+                    return new ResponseEntity<>(line2.toString(), HttpStatus.OK);
+                }
+            }
+           return null;
         } else {
             return null;
         }
